# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  whatsapp?: string | null;
  course: string;
  year: number;
  subjects: string[];
  availability: string[];
}

export interface UserProfileInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @nullable */
  whatsapp?: string | null;
  /** @minLength 1 */
  course: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  year: number;
  subjects: string[];
  availability: string[];
}

export type StudyRequestLocation = typeof StudyRequestLocation[keyof typeof StudyRequestLocation];


export const StudyRequestLocation = {
  campus: 'campus',
  online: 'online',
} as const;

export interface StudyRequest {
  id: number;
  subject: string;
  topic: string;
  preferredDate: string;
  preferredTime: string;
  location: StudyRequestLocation;
  note: string;
  creatorName: string;
  creatorCourse: string;
  creatorYear: number;
  joinedCount: number;
  hasJoined: boolean;
  isOwner: boolean;
  /** @nullable */
  contactInfo: string | null;
}

export type StudyRequestInputLocation = typeof StudyRequestInputLocation[keyof typeof StudyRequestInputLocation];


export const StudyRequestInputLocation = {
  campus: 'campus',
  online: 'online',
} as const;

export interface StudyRequestInput {
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  topic: string;
  preferredDate: string;
  /** @minLength 1 */
  preferredTime: string;
  location: StudyRequestInputLocation;
  /** @minLength 1 */
  note: string;
}

export interface SubjectCount {
  label: string;
  count: number;
}

export interface Analytics {
  totalRequests: number;
  totalMembers: number;
  activeToday: number;
  popularSubjects: SubjectCount[];
  busySlots: SubjectCount[];
}

export type ListStudyRequestsParams = {
search?: string;
subject?: string;
date?: string;
location?: ListStudyRequestsLocation;
};

export type ListStudyRequestsLocation = typeof ListStudyRequestsLocation[keyof typeof ListStudyRequestsLocation];


export const ListStudyRequestsLocation = {
  campus: 'campus',
  online: 'online',
} as const;