# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubjectCount } from './subjectCount';

export interface Analytics {
  totalRequests: number;
  totalMembers: number;
  activeToday: number;
  popularSubjects: SubjectCount[];
  busySlots: SubjectCount[];
}