# lib/api-zod/src/generated/types/studyRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudyRequestLocation } from './studyRequestLocation';

export interface StudyRequest {
  id: number;
  subject: string;
  topic: string;
  preferredDate: Date;
  preferredTime: string;
  location: StudyRequestLocation;
  note: string;
  creatorName: string;
  creatorCourse: string;
  creatorYear: number;
  joinedCount: number;
  hasJoined: boolean;
  isOwner: boolean;
  /** @nullable */
  contactInfo: string | null;
}