# lib/api-zod/src/generated/types/studyRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudyRequestInputLocation } from './studyRequestInputLocation';

export interface StudyRequestInput {
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  topic: string;
  preferredDate: Date;
  /** @minLength 1 */
  preferredTime: string;
  location: StudyRequestInputLocation;
  /** @minLength 1 */
  note: string;
}