# lib/api-zod/src/generated/types/userProfileInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @nullable */
  whatsapp?: string | null;
  /** @minLength 1 */
  course: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  year: number;
  subjects: string[];
  availability: string[];
}